Handle PR comments and add some tests to prevent regressions

+++ app/controllers/api/v1/articles_controller.rb
@@ -24,6 +24,10 @@ class Api::V1::ArticlesController < Api::V1::BaseController
   def destroy
     article = Article.find_by(id: params[:id], user: current_user)
 
+    unless article
+      return render json: { error: 'article not found' }, status: :not_found
+    end
+
     if article.destroy
       head :no_content
     else
